pyproject.toml
[project]
name = "careerconversationai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.39.0",
    "openai>=1.98.0",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
